move utilities to utils folder and icons to assets folder, replace arrayToIdMap with lodash
