+++ packages/app/src/actions/dimensions.js
@@ -1,11 +1,12 @@
-import { arrayToIdMap, sortArray } from '../util';
+import keyBy from 'lodash-es/keyBy';
+import { sortArray } from '../utils/util';
 import { SET_DIMENSIONS } from '../reducers/dimensions';
 import { apiFetchDimensions } from '../api/dimensions';
 import { sGetDisplayNameProperty } from '../reducers/settings';
 
 export const acSetDimensions = dimensions => ({
     type: SET_DIMENSIONS,
-    value: arrayToIdMap(sortArray(dimensions, 'name')),
+    value: keyBy(sortArray(dimensions, 'name'), 'id'),
 });
 
 export const tSetDimensions = () => async (dispatch, getState) => {
