+++ packages/app/src/api/visualization.js
@@ -1,5 +1,5 @@
 import { getInstance } from 'd2';
-import { getFieldsStringByType } from '../fields';
+import { getFieldsStringByType } from '../utils/fields';
 
 export const apiFetchVisualization = (type, id) =>
     getInstance().then(d2 =>
