+++ packages/app/src/components/App.js
@@ -13,8 +13,8 @@ import BlankCanvas from './Visualization/BlankCanvas';
 import Layout from './Layout/Layout';
 import * as fromReducers from '../reducers';
 import * as fromActions from '../actions';
-import history from '../history';
-import defaultMetadata from '../metadata';
+import history from '../utils/history';
+import defaultMetadata from '../utils/metadata';
 
 import './App.css';
 import { sGetUi } from '../reducers/ui';
