+++ packages/app/src/components/Layout/Chip.js
@@ -4,12 +4,12 @@ import i18n from '@dhis2/d2-i18n';
 
 import Menu from './Menu';
 import Tooltip from './Tooltip';
-import { setDataTransfer } from '../../dnd';
+import { setDataTransfer } from '../../utils/dnd';
 import { sGetDimensions } from '../../reducers/dimensions';
 import { sGetUiItems } from '../../reducers/ui';
 import { styles } from './styles/Chip.style';
-import { FIXED_DIMENSIONS } from '../../fixedDimensions';
-import { GenericDimensionIcon } from '../../icons';
+import { FIXED_DIMENSIONS } from '../../utils/fixedDimensions';
+import { GenericDimensionIcon } from '../../assets/icons';
 
 const TOOLTIP_ENTER_DELAY = 500;
