+++ packages/app/src/components/Layout/styles/Tooltip.style.js
@@ -1,4 +1,4 @@
-import { colors } from '../../../colors';
+import { colors } from '../../../utils/colors';
 
 export const styles = {
     tooltip: {
