+++ packages/app/src/components/UpdateButton/styles/UpdateButton.style.js
@@ -1,4 +1,4 @@
-import { colors } from '../../../colors';
+import { colors } from '../../../utils/colors';
 
 export default {
     backgroundColor: colors.accentPrimaryDark,
