+++ packages/app/src/components/Visualization/__tests__/Visualization.spec.js
@@ -3,7 +3,7 @@ import { shallow } from 'enzyme';
 import * as chartsApi from 'd2-charts-api';
 import { Visualization } from '../Visualization';
 import BlankCanvas from '../BlankCanvas';
-import * as options from '../../../options';
+import * as options from '../../../utils/options';
 
 const metaDataMock = ['a', 'b'];
 class MockAnalyticsResponse {
