+++ packages/app/src/reducers/__tests__/ui.spec.js
@@ -15,7 +15,7 @@ import reducer, {
     SET_UI_ACTIVE_MODAL_DIALOG,
     CLEAR_UI,
 } from '../ui';
-import { AXIS_NAMES } from '../../layout';
+import { AXIS_NAMES } from '../../utils/layout';
 
 const [COLUMNS, ROWS, FILTERS] = AXIS_NAMES;
