+++ packages/app/src/reducers/dimensions.js
@@ -1,4 +1,4 @@
-import { FIXED_DIMENSIONS as DEFAULT_DIMENSIONS } from '../fixedDimensions';
+import { FIXED_DIMENSIONS as DEFAULT_DIMENSIONS } from '../utils/fixedDimensions';
 
 export const SET_DIMENSIONS = 'SET_DIMENSIONS';
 export const SET_SELECTED_DIMENSION = 'SET_SELECTED_DIMENSION';
