+++ packages/app/src/utils/current.js
